requirements.txt
Flask==2.3.3
openai>=1.6.0
python-dotenv==1.0.0
Werkzeug==2.3.7
httpx>=0.24.1
gunicorn==20.1.0
